Finished saved status check.
Added modal for confirmation while closing the app with changes on the guides.
Saving a guide sets its status to "no changes", but it happens regardless of whether the user ends up saving the file or not. Implementing this (check whether it actually gets saved locally) can't be done in the rendering side of the app it seems